+++ app/admin/dashboard/page.tsx
@@ -1,23 +1,22 @@
-"use client";
-
 import React from "react";
-import { useRouter } from "next/navigation";
-
-import AuthorizationState from "@/app/_lib/client/authorization";
+import dynamic from "next/dynamic";
+import { Metadata } from "next";
 
 import "@/public/static/css/pages.css";
 
-export default function Home(): React.JSX.Element {
-  if (!AuthorizationState.instance.loggedIn) {
-    const router = useRouter();
-    router.push("/admin");
-  }
+export const metadata: Metadata = {
+  title: "Administrator dashboard",
+};
 
+// https://nextjs.org/docs/pages/building-your-application/optimizing/lazy-loading#with-no-ssr
+const DynamicRedirect = dynamic(() => import("./Redirect"), { ssr: false });
+
+export default function Home(): React.JSX.Element {
   return (
     <div className="page-section">
       <div className="intro-section">
         <div className="header-1" style={{ color: "black" }}>
-          <h1>Welcome <u>{AuthorizationState.instance.username}</u>!</h1>
+          <DynamicRedirect />
         </div>
       </div>
     </div>
